Version 0.001
This was the first version of the suggestion. There has been a number of comments based on this first version. One of the main points was that this version didn’t consider colorblind people. The first draft needs therefore to be remade so it takes colorblind people into consideration. The solution should not have close hues for close orders of magnitude of bitcoin price.
It has also been suggested that add additional markers could help, such as 
BTC 4,321.000,543,21
(is BTC 4,321 and 543.21 bits)
 
It was also suggested to make use of warmer colors for the higher values and colder colors for the lower values. This has of course to be incorporated while still taking the colorblind people into consideration.
The most critical comments pointed out that text is not good legibility in colors, monochrome high contrast being the best and that UX colors have expected meaning. 
The next version needs therefore to take all of these points into consideration.